+++ Model/punto.cpp
@@ -1,13 +1,13 @@
 #include "punto.h"
 #include <math.h>
 
-Punto::Punto(unsigned int a, unsigned int b) : x(a), y(b) {}
+Punto::Punto(double a, double b) : x(a), y(b) {}
 
-unsigned int Punto::getX() const {
+double Punto::getX() const {
     return x;
 }
 
-unsigned int Punto::getY() const {
+double Punto::getY() const {
     return y;
 }
 
@@ -20,28 +20,28 @@ void Punto::setPunto(const Punto& p) {
     y = p.getY();
 }
 
-void Punto::setX(unsigned int i) {
+void Punto::setX(double i) {
     x = i;
 }
 
-void Punto::setY(unsigned int i) {
+void Punto::setY(double i) {
     y = i;
 }
 
-unsigned int Punto::distanza(const Punto& p) const {
+double Punto::distanza(const Punto& p) const {
 
-    unsigned int a = pow(p.getX() - x, 2);
-    unsigned int b = pow(p.getY() - y, 2);
-    unsigned int dist = sqrt(a+b);
+    double a = pow(p.getX() - x, 2);
+    double b = pow(p.getY() - y, 2);
+    double dist = sqrt(a+b);
     return dist;
 
 
 }
 
 double Punto::distanzaDuePunti(const Punto& a, const Punto& b) {
-    unsigned int v = pow(b.getX() - a.getX(), 2);
-    unsigned int z = pow(b.getY() - a.getY(), 2);
-    unsigned int dist = sqrt(v+z);
+    double v = pow(b.getX() - a.getX(), 2);
+    double z = pow(b.getY() - a.getY(), 2);
+    double dist = sqrt(v+z);
     return dist;
 
 }
